requirements.txt
anthropic
mcp
openai
python-dotenv
langchain
langgraph
langchain-anthropic
langchain-openai
langchain-mcp-adapters
colorama